Ignore key fragments in primary keyring processing.
Revocation certificates will require special treatment, if it is in fact
the keyserver's responsibility to match the key ID and add it to the
keyring.

That will require special case handling in the /pks/add handler; we'd
have to look up the key ID and do the merging for the client.

At least this avoids panicking.